[proliferate] split review defaults section
